Update
The Sol-Toolbox-For-Education project is an educational repository containing a variety of Solidity contract examples and utility libraries designed for developers to learn and explore the Ethereum ecosystem. The main goal of this project is to provide a comprehensive set of resources to help developers understand, learn and experiment with different functionalities and concepts related to Ethereum smart contracts.